refactor: extract Flag 6 LLM backend from filter_flags.py (#559)
filter_flags.py was 703L — 60% of it was Flag 6 (LLM relevance scoring):
cache management, model loading, reranker/LLM backends, batch streaming.
The basic flags (1-5) and protection are independent concerns that don't
need to change when the LLM backend evolves.

Split into:
- filter_flags.py (291L): flags 1-5, protection, helpers
- filter_flags_llm.py (475L): Flag 6 cache, backends, streaming

_text_has_concept_groups duplicated in LLM module to avoid circular import
(filter_flags re-exports LLM functions for backward compatibility).

Also:
- dvc.yaml: added filter_flags_llm.py as dep for extend/filter stages
- tests: monkeypatch targets updated to filter_flags_llm module
- test_script_hygiene: added exit-criterion tests for the split
